Update to djikstraAlg read description
djikstraAlg does not work as a driver function at the moment as I was not able to come up with a way to set up distance, cost, and time as parameters for the method. Right now it calculates based on cost. It is not fully functional at the moment as it look for the edge with the cheapest value instead of looking for a path to the end. Must be updated to explore entire graph. Need new way to mark the graph so that there is an already visited, currently visiting and not yet explored. Perhaps a Queue implementation. Currently using a hashmap I was unable to fully grasp djikstra's algorithm. However I am close to getting it. Currently trying to find a way to make this a recursive method in order to explore the entire graph.